% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHSMChart.R
\name{PHSMChart}
\alias{PHSMChart}
\title{PHSMChart
This function will build a chart showing the epicurve against PHSM measures}
\usage{
PHSMChart(SeverityExcel, country)
}
\arguments{
\item{SeverityExcel}{path to the severity excel file (build by PHSM team)}

\item{country}{the country to build the chart for}
}
\description{
PHSMChart
This function will build a chart showing the epicurve against PHSM measures
}
